lighthouse_v1/pyproject.toml
[project]
name = "linux-lighthouse"
version = "0.1.0"
description = "Terminal-native Linux decision and guidance system"
requires-python = ">=3.11"
dependencies = ["pyyaml"]

[project.scripts]
lighthouse = "lighthouse.main:main"

